LayoutTests/http/tests/security/xss-DENIED-non-shadowable-propterty-with-iframe-proto-expected.txt


Tests that making other frame window a prototype doesn't expose that window properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS __proto__ = targetWindow threw exception TypeError: Cannot set prototype of this object.
PASS: location === originalLocation should be 'true' and is.
PASS: this.location === originalLocation should be 'true' and is.
PASS: successfullyParsed should be 'true' and is.

TEST COMPLETE
